requirements.txt
amqp==5.0.5
asgiref==3.3.1
auditlog3==1.0.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-3-jet
django-utils-six==2.0
django-crontab==0.7.1
django-currentuser==0.5.2
idna==2.10
kombu==5.0.2
plotly==4.12.0
prompt-toolkit==3.0.16
pytz==2020.4
requests==2.25.0
schedule==1.0.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
